Corrected a typo in the sonata-demo example service descriptor.
Corrected typo in the sonata-demo example service descriptor as mentioned in GitHub issue 96. That is, the service descriptor contained an incorrect connectivity graph in the virtual_links section. In line 80 it is defined "vnf_firewall:output" but should be "vnf_tcpdump:output". Subsequently, the versio and the MD5 sum of the descriptor changed, such that the package descriptor and the package had to be updated.